+++ src/NadekoBot/Modules/Games/Common/Hangman/HangmanGame.cs
@@ -73,7 +73,6 @@ namespace NadekoBot.Modules.Games.Hangman
             return new(output);
         }
 
-        // todo lock
         public State Guess(string guess)
         {
             if (CurrentPhase != Phase.Running)